Moved ViewModelbaseTests to the new ModelBaseTests
